Added a dict-like items() function and fixed __iter__() for dicts
The list iterator gives back iterators to traversers of subtrees in that list. In case of a dict, it would be nice to get back traverses of subtrees for dictionaries as well, my change in __iter__() fixes just that.

I also added a function items(). Similar to this function for a dict, it gives back a list of subtrees for a list, and tuples like this (dict_key, subtree_for_key). As far as I have tested these two additions work like a charm. In order to run my program I needed both, and now it's really awesome to use traversify! 

If you think these changes are ok, I'd be glad if you could merge them into a newer version :)